feat: create mention_users global skill
Move user mention functionality from generation.ts to a new global skill.
The skill provides context-aware instructions for mentioning users,
with different behavior for Slack/Teams vs web origins.

This change:
- Creates mention_users.ts global skill with fetchInstructions
- Removes mention section from constructGuidelinesSection
- Fixes listByAgentConfiguration calls to pass object parameter
- Updates tests to remove mention-specific test cases
