+++ front/components/resources/resources_icons.tsx
@@ -22,6 +22,7 @@ import {
   ActionSpeakIcon,
   ActionTableIcon,
   ActionTimeIcon,
+  ActionUserIcon,
   AsanaLogo,
   AshbyLogo,
   Avatar as SparkleAvatar,
@@ -111,6 +112,7 @@ export const InternalActionIcons = {
   ActionSpeakIcon,
   ActionTableIcon,
   ActionTimeIcon,
+  ActionUserIcon,
   AsanaLogo,
   AshbyLogo,
   ToolsIcon,
