+++ front/lib/actions/mcp_internal_actions/servers/index.ts
@@ -38,6 +38,7 @@ import { default as includeDataServer } from "@app/lib/api/actions/servers/inclu
 import { default as interactiveContentServer } from "@app/lib/api/actions/servers/interactive_content";
 import { default as jiraServer } from "@app/lib/api/actions/servers/jira";
 import { default as jitTestingServer } from "@app/lib/api/actions/servers/jit_testing";
+import { default as mentionUsersServer } from "@app/lib/api/actions/servers/mention_users";
 import { default as microsoftDriveServer } from "@app/lib/api/actions/servers/microsoft_drive";
 import { default as microsoftExcelServer } from "@app/lib/api/actions/servers/microsoft_excel";
 import { default as microsoftTeamsServer } from "@app/lib/api/actions/servers/microsoft_teams";
@@ -241,6 +242,8 @@ export async function getInternalMCPServer(
       return statuspageServer(auth, agentLoopContext);
     case "sandbox":
       return sandboxServer(auth, agentLoopContext);
+    case "mention_users":
+      return mentionUsersServer(auth, agentLoopContext);
     default:
       assertNever(internalMCPServerName);
   }
